+++ src/spyglass/position/v1/position_dlc_selection.py
@@ -57,6 +57,14 @@ class DLCPosV1(SpyglassMixin, dj.Computed):
     """
 
     def make(self, key):
+        """Populate the table with the combined position information.
+
+        1. Fetches position and orientation data from the DLCCentroid and
+        DLCOrientation tables.
+        2. Creates NWB objects for position, orientation, and velocity.
+        3. Generates an AnalysisNwbfile and adds the NWB objects to it.
+        4. Inserts the key into the table, and the PositionOutput Merge table.
+        """
         orig_key = copy.deepcopy(key)
         # Add to Analysis NWB file
         AnalysisNwbfile()._creation_times["pre_create_time"] = time()
@@ -149,7 +157,8 @@ class DLCPosV1(SpyglassMixin, dj.Computed):
         )
         AnalysisNwbfile().log(key, table=self.full_table_name)
 
-    def fetch1_dataframe(self):
+    def fetch1_dataframe(self) -> pd.DataFrame:
+        """Return the position data as a DataFrame."""
         nwb_data = self.fetch_nwb()[0]
         index = pd.Index(
             np.asarray(nwb_data["position"].get_spatial_series().timestamps),
@@ -188,11 +197,13 @@ class DLCPosV1(SpyglassMixin, dj.Computed):
         )
 
     def fetch_nwb(self, **kwargs):
+        """Fetch the NWB file."""
         attrs = [a for a in self.heading.names if not a == "pose_eval_result"]
         return super().fetch_nwb(*attrs, **kwargs)
 
     @classmethod
     def evaluate_pose_estimation(cls, key):
+        """Evaluate the pose estimation."""
         likelihood_thresh = []
 
         valid_fields = DLCSmoothInterpCohort.BodyPart().heading.names
@@ -272,6 +283,7 @@ class DLCPosVideoParams(SpyglassMixin, dj.Manual):
 
     @classmethod
     def insert_default(cls):
+        """Insert the default parameters."""
         params = {
             "percent_frames": 1,
             "incl_likelihood": True,
@@ -287,6 +299,7 @@ class DLCPosVideoParams(SpyglassMixin, dj.Manual):
 
     @classmethod
     def get_default(cls):
+        """Return the default parameters."""
         query = cls & {"dlc_pos_video_params_name": "default"}
         if not len(query) > 0:
             cls().insert_default()
@@ -318,6 +331,14 @@ class DLCPosVideo(SpyglassMixin, dj.Computed):
     """
 
     def make(self, key):
+        """Populate the DLCPosVideo table.
+
+        1. Fetches parameters from the DLCPosVideoParams table.
+        2. Fetches position interval name from epoch name.
+        3. Fetches pose estimation data and video information.
+        4. Fetches centroid and likelihood data for each bodypart.
+        5. Calls make_video to create the video with the above data.
+        """
         M_TO_CM = 100
 
         params = (DLCPosVideoParams & key).fetch1("params")