+++ src/spyglass/spikesorting/v0/figurl_views/SpikeSortingRecordingView.py
@@ -1,4 +1,3 @@
-import os
 from typing import List, Union
 
 import datajoint as dj
@@ -26,6 +25,11 @@ class SpikeSortingRecordingView(SpyglassMixin, dj.Computed):
     """
 
     def make(self, key):
+        """Populates SpikeSortingRecordingView.
+
+        Fetches the recording from SpikeSortingRecording and extracts traces
+        and electrode geometry.
+        """
         # Get the SpikeSortingRecording row
         rec = (SpikeSortingRecording & key).fetch1()
         nwb_file_name = rec["nwb_file_name"]
@@ -66,6 +70,7 @@ class SpikeSortingRecordingView(SpyglassMixin, dj.Computed):
 
 
 def create_electrode_geometry(recording: si.BaseRecording):
+    """Create a figure for the electrode geometry of a recording."""
     channel_locations = {
         str(channel_id): location.astype(np.float32)
         for location, channel_id in zip(
@@ -81,6 +86,7 @@ def create_mountain_layout(
     label: Union[str, None] = None,
     sorting_curation_uri: Union[str, None] = None,
 ) -> Figure:
+    """Create a figure for a mountain layout of multiple figures"""
     if label is None:
         label = "SpikeSortingView"