+++ src/spyglass/spikesorting/v0/figurl_views/SpikeSortingView.py
@@ -24,6 +24,21 @@ class SpikeSortingView(SpyglassMixin, dj.Computed):
     """
 
     def make(self, key):
+        """Populates SpikeSortingView.
+
+        1. Fetches...
+            - the recording from SpikeSortingRecording
+            - the sorting from SpikeSorting
+        2. Loads each with spikeinterface and passes to SpikeSortingView from
+            sortingview package.
+        3. Creates...
+            - Summary
+            - Autocorrelograms
+            - Average waveforms
+            - Spike amplitudes
+            - Electrode geometry
+        4. Creates a mountain layout with the above figures and generates a URL.
+        """
         recording_record = (
             SpikeSortingRecording & {"recording_id": key["recording_id"]}
         ).fetch1()