+++ src/spyglass/utils/dj_graph.py
@@ -939,6 +939,7 @@ class TableChain(RestrGraph):
 
     @property
     def path_str(self) -> str:
+        """Return string representation of path: parent -> {links} -> child."""
         if not self.path:
             return "No link"
         return self._link_symbol.join([self._camel(t) for t in self.path])
@@ -981,6 +982,7 @@ class TableChain(RestrGraph):
     # ---------------------------- Graph Traversal ----------------------------
 
     def cascade_search(self) -> None:
+        """Cascade restriction through graph to search for applicable table."""
         if self.cascaded:
             return
         restriction, restr_attrs = self._get_find_restr(self.leaf)
@@ -1033,6 +1035,7 @@ class TableChain(RestrGraph):
         limit: int = 100,
         **kwargs,
     ):
+        """Search parents/children for a match of the provided restriction."""
         if (
             self.found_restr
             or not table
@@ -1135,6 +1138,7 @@ class TableChain(RestrGraph):
     def cascade(
         self, restriction: str = None, direction: Direction = None, **kwargs
     ):
+        """Cascade restriction up or down the chain."""
         if not self.has_link:
             return